move record reader to entry:io
