+++ reductstore/src/storage/entry/io.rs
@@ -1,0 +1,5 @@
+// Copyright 2024 ReductSoftware UG
+// Licensed under the Business Source License 1.1
+
+pub(crate) mod record_reader;
+pub(crate) mod record_writer;
