+++ reductstore/src/storage/entry/write_record.rs
@@ -2,7 +2,7 @@
 // Licensed under the Business Source License 1.1
 
 use crate::storage::block_manager::{BlockRef, RecordTx};
-use crate::storage::entry::record_writer::WriteRecordContent;
+use crate::storage::entry::io::record_writer::WriteRecordContent;
 use crate::storage::entry::{Entry, RecordType, RecordWriter};
 use crate::storage::proto::{record, us_to_ts, Record};
 use log::debug;
