+++ reductstore/src/storage/query/limited.rs
@@ -2,11 +2,11 @@
 // Licensed under the Business Source License 1.1
 
 use crate::storage::block_manager::BlockManager;
-use crate::storage::bucket::RecordReader;
 use crate::storage::query::base::{Query, QueryOptions};
 use crate::storage::query::historical::HistoricalQuery;
 use async_trait::async_trait;
 
+use crate::storage::entry::RecordReader;
 use reduct_base::error::{ErrorCode, ReductError};
 use std::sync::Arc;
 use tokio::sync::RwLock;
